requirements.txt
sqlalchemy
fastapi
python-dotenv
botocore
boto3
psycopg_pool
langchain_openai
langchain
langgraph
psycopg
langgraph-checkpoint-postgres
langchain_ollama
psycopg2-binary
qdrant_client
langchain_qdrant
qdrant_client
opencv-python
pdf2image
PyMuPDF